refactoring existing functionality to new class
