+++ BridgeCareApp/BridgeCare/UtilityFunctions.cs
@@ -18,27 +18,6 @@ namespace BridgeCare.Utility
             }
             return resultsDataTable;
         }
-
-        public static bool IsDigitsOnly(string str)
-        {
-            foreach (char c in str)
-            {
-                if (c < '0' || c > '9')
-                    return false;
-            }
-
-            return true;
-        }
-
-        //check for year
-        public static bool IsIamYear(string str)
-        {
-            if (IsDigitsOnly(str) && str.Length == 4)
-            {
-                return true;
-            }
-
-            return false;
-        }
     }
+
 }
